Исправила ошибку в задаче38

+++ Task38/Program.cs
@@ -16,7 +16,7 @@ PrintArray(array);
 double max = array[0];
 double min = array[0];
 
-for (double i = 0; i < array.Length; i++)
+for (int i = 0; i < array.Length; i++)
 {
     if (array[i] > max)
     {
